+++ src/main/java/devsprint/omuk/recipe/domain/MealTime.java
@@ -1,4 +1,4 @@
-package devsprint.omuk.domain;
+package devsprint.omuk.recipe.domain;
 
 public enum MealTime {
     BREAKFAST, LUNCH, DINNER, LATE_NIGHT
