+++ src/main/java/devsprint/omuk/recipe/domain/Season.java
@@ -1,4 +1,4 @@
-package devsprint.omuk.domain;
+package devsprint.omuk.recipe.domain;
 
 public enum Season {
     SPRING, SUMMER, FALL, WINTER
